Simplified timing

+++ Libraries/Bro/Library/Include/Bro/Bro.h
@@ -11,6 +11,6 @@ void broTerminate();
 void broRequestTermination();
 void broSwapBuffers();
 bool broIsVisible();
-uint64_t broGetTime();
+double broGetTime();
 
 #endif
