+++ User/MainFlow/State.h
@@ -5,7 +5,7 @@ namespace MainFlow {
   class State {
   public:
     virtual void enter() { }
-    virtual void update(float deltaTime) = 0;
+    virtual void update(double deltaTime) = 0;
     virtual void exit() { }
     virtual State* checkTransition() = 0;
     virtual ~State() { }
